% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arrayconversion}
\alias{arrayconversion}
\title{Convert an array of measurements of length}
\usage{
arrayconversion(value, oldunit, newunit = "meter")
}
\arguments{
\item{value}{Array of values to convert}

\item{oldunit}{Unit of the value to convert}

\item{newunit}{Unit of return value, default is meters}
}
\value{
Converted array of values
}
\description{
Convert an array of measurements of length
}
\examples{
arrayconversion(c(5,10,15),"yd","mi")
arrayconversion(c(5,10,15),"yard")
}
