% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arraystrconversion}
\alias{arraystrconversion}
\title{Convert an array of strings with measurements of length}
\usage{
arraystrconversion(old, newunit = "meter", returnstring = FALSE)
}
\arguments{
\item{old}{Array of strings with values with unit at the end}

\item{newunit}{Unit of return value, default is meters}

\item{returnstring}{TRUE:return is string with unit name at the end, FALSE: return is numeric data}
}
\value{
Converted array of values
}
\description{
Convert an array of strings with measurements of length
}
\examples{
arraystrconversion(c("5yd","6ft"),"mi")
arraystrconversion(c("10miles","3foot","2chain"),"yard",TRUE)
}
