% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise6.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit}
\usage{
C_to_F(c_temp)
}
\arguments{
\item{c_temp}{The Celsius temperature to be converted}
}
\value{
The converted Fahrenheit temperature
}
\description{
Celsius to Fahrenheit
}
\examples{
C_to_F(100)
C_to_F(0)

}
