% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise6.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin}
\usage{
C_to_K(c_temp)
}
\arguments{
\item{c_temp}{The Celsius temperature to be converted}
}
\value{
The converted Kelvin temperature
}
\description{
Celsius to Kelvin
}
