% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{concatenate_strings}
\alias{concatenate_strings}
\title{Concatenate Two Strings}
\usage{
concatenate_strings(str1, str2)
}
\arguments{
\item{str1}{A character string.}

\item{str2}{A character string.}
}
\value{
single character string with str1 and str2 concatenated with a space in between.
}
\description{
This function takes two strings and concatenates them with a space in between.
}
\examples{
concatenate_strings("Hello", "World")
}
