% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise6.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius}
\usage{
F_to_C(f_temp)
}
\arguments{
\item{f_temp}{The Fahrenheit temperature to be converted}
}
\value{
The converted Celsius temperature
}
\description{
Fahrenheit to Celsius
}
\examples{
F_to_C(100)
F_to_C(0)

}
