% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise6.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin}
\usage{
F_to_K(f_temp)
}
\arguments{
\item{f_temp}{The Fahrenheit temperature to be converted}
}
\value{
The converted Kelvin temperature
}
\description{
Fahrenheit to Kelvin
}
\examples{
F_to_K(100)
F_to_K(0)

}
