% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise6.R
\name{K_to_C}
\alias{K_to_C}
\title{Kelvin to Celsius}
\usage{
K_to_C(k_temp)
}
\arguments{
\item{k_temp}{The Kelvin temperature to be converted}
}
\value{
The converted Celsius temperature
}
\description{
Kelvin to Celsius
}
