% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise6.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit}
\usage{
K_to_F(k_temp)
}
\arguments{
\item{k_temp}{The Kelvin temperature to be converted}
}
\value{
The converted Fahrenheit temperature
}
\description{
Kelvin to Fahrenheit
}
