% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{read_data_file}
\alias{read_data_file}
\title{Funktion zum Einlesen von Daten aus verschiedenen Dateiformaten.}
\usage{
read_data_file(file_path)
}
\arguments{
\item{file_path}{Pfad zur Datei.}
}
\value{
Gibt die eingelesen Tabelle zurück.
}
\description{
Nimmt den Pfad der Datei entgegen und liest die Datei ein.
}
\examples{
# Basic example
table <- read_data_file("C:/Users/name/Desktop/data/Data.csv")
reads the data from the specified path.
}
