% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a String}
\usage{
reverse_string(str)
}
\arguments{
\item{str}{A character string.}
}
\value{
A character string with the characters of str in reverse order.
}
\description{
This function takes a string and returns it reversed.
}
\examples{
reverse_string("Hello")
}
