% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strconversion}
\alias{strconversion}
\title{Convert a string with measurements of length}
\usage{
strconversion(old, newunit = "meter", returnstring = FALSE)
}
\arguments{
\item{old}{String of values with unit at the end}

\item{newunit}{Unit of return value, default is meters}

\item{returnstring}{TRUE:return is string with unit name at the end, FALSE: return is numeric data}
}
\value{
Converted value
}
\description{
Convert a string with measurements of length
}
\examples{
strconversion("5yd","mi")
strconversion("10miles","yard",TRUE)
}
