% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{write_data_file}
\alias{write_data_file}
\title{Funktion zum schreiben von Daten in verschiedenen Dateiformaten.}
\usage{
write_data_file(data, file_path)
}
\arguments{
\item{data}{Tabelle die geschrieben werden soll.}

\item{file_path}{Pfad zur Datei.}
}
\value{
erstellt eine Date bei dem angegebenen Pfad, im angegebenen Format.
}
\description{
Nimmt einen Pfad und eine Tabelle entgegen und schreibt die Tabelle in die Datei.
}
\examples{
# Basic example
write_data_file(data = data, file_path = "C:/Users/name/Desktop/data/Data.csv")
writes the data table to the specified path in the specified format.
}
