% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conversion}
\alias{conversion}
\title{Convert multiple measurements of length}
\usage{
conversion(value, oldunit, newunit = "meter")
}
\arguments{
\item{value}{Value to convert}

\item{oldunit}{Unit of the value to convert}

\item{newunit}{Unit of return value, default is meters}
}
\value{
Converted value
}
\description{
Possible units: "twip", "thou", "inch", "hand", "foot", "yard", "chain", "furlong", "mile", "league", "millimeter", "centimeter", "meter", "kilometer" or "twip", "th", "in", "hh", "ft", "yd", "ch", "fur", "mi", "lea", "mm", "cm", "m", "km"
}
\examples{
conversion(5,"yd","mi")
conversion(10,"yard")
}
