% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{strconversion}
\alias{strconversion}
\title{Convert a string with measurements of length}
\usage{
strconversion(old, newunit = "meter", returnstring = FALSE)
}
\arguments{
\item{old}{String of values with unit at the end}

\item{newunit}{Unit of return value, default is meters}

\item{returnstring}{TRUE:return is string with unit name at the end, FALSE: return is numeric data}
}
\value{
Converted value
}
\description{
Possible units: "twip", "thou", "inch", "hand", "foot", "yard", "chain", "furlong", "mile", "league", "millimeter", "centimeter", "meter", "kilometer" or "twip", "th", "in", "hh", "ft", "yd", "ch", "fur", "mi", "lea", "mm", "cm", "m", "km"
Plurals "-s" and spaces in the string will be ignored
}
\examples{
strconversion("5yd","mi")
strconversion("10miles","yard",TRUE)
}
